Test TryInv, remove unnecessary reductions
When calling TryInv or Inv, we already know the number m/n is in reduced form,
so we also know n/m will be. Moreover, when multiplying two numbers a/b and
c/d already in reduced form, if we reduce a and d by their GCD and b and c by
their GCD, we get the product (a*c)/(b*d) already in reduced form as well.
Therefore, it is not necessary to compute the GCDs again, and so we can skip
a step of reducing these numbers.